Make separate function for generating regular noun forms in German
