+++ packages/yoastseo/spec/morphology/german/getGermanFormsSpec.js
@@ -1,7 +1,6 @@
 import { getForms } from "../../../src/morphology/german/getForms";
 import getMorphologyData from "../../specHelpers/getMorphologyData";
 
-
 const morphologyDataDE = getMorphologyData( "de" ).de;
 
 describe( "Test for creating forms for German words", () => {
@@ -38,219 +37,6 @@ describe( "Test for creating forms for German words", () => {
 			"gestudentet",
 			// Participle.
 			"student" ] );
-	} );
-
-	it( "creates forms for words that don't get get an -s noun suffix (but all other regular suffixes)", () => {
-		expect( getForms( "spieß", morphologyDataDE ) ).toEqual( [
-			"spieß",
-			"spieße",
-			"spießen",
-			"spießens",
-			"spießer",
-			"spießern",
-			"spießers",
-			"spießes",
-			"spießem",
-			"spießere",
-			"spießerem",
-			"spießeren",
-			"spießerer",
-			"spießeres",
-			"spießste",
-			"spießstem",
-			"spießsten",
-			"spießster",
-			"spießstes",
-			"spießt",
-			"spießest",
-			"spießet",
-			"spießte",
-			"spießtet",
-			"spießten",
-			"spießtest",
-			"spießete",
-			"spießetet",
-			"spießeten",
-			"spießetest",
-			"spießend",
-			"gespießt",
-		] );
-	} );
-
-	it( "creates forms for words which get -n and -ns noun suffixes in addition to all regular noun suffixes", () => {
-		expect( getForms( "winkel", morphologyDataDE ) ).toEqual( [
-			// Regular noun suffixes
-			"winkel",
-			"winkele",
-			"winkelen",
-			"winkelens",
-			"winkeler",
-			"winkelern",
-			"winkelers",
-			"winkeles",
-			"winkels",
-			// Added noun suffixes
-			"winkeln",
-			"winkelns",
-			// Adjective suffixes
-			"winkelem",
-			"winkelere",
-			"winkelerem",
-			"winkeleren",
-			"winkelerer",
-			"winkeleres",
-			"winkelste",
-			"winkelstem",
-			"winkelsten",
-			"winkelster",
-			"winkelstes",
-			"winkelt",
-			"winkelst",
-			"winkelest",
-			"winkelet",
-			"winkelte",
-			"winkeltet",
-			"winkelten",
-			"winkeltest",
-			"winkelete",
-			"winkeletet",
-			"winkeleten",
-			"winkeletest",
-			"winkelend",
-			// Participle.
-			"gewinkelt",
-		] );
-	} );
-
-	it( "creates forms for words which get the suffix -nen in addition to all other regular noun suffixes", () => {
-		expect( getForms( "ärztin", morphologyDataDE ) ).toEqual( [
-			// Regular noun suffixes
-			"ärztin",
-			"ärztine",
-			"ärztinen",
-			"ärztinens",
-			"ärztiner",
-			"ärztinern",
-			"ärztiners",
-			"ärztines",
-			"ärztins",
-			// Additional noun suffixes
-			"ärztinnen",
-			// Adjective suffixes
-			"ärztinem",
-			"ärztinere",
-			"ärztinerem",
-			"ärztineren",
-			"ärztinerer",
-			"ärztineres",
-			"ärztinste",
-			"ärztinstem",
-			"ärztinsten",
-			"ärztinster",
-			"ärztinstes",
-			"ärztint",
-			"ärztinst",
-			"ärztinest",
-			"ärztinet",
-			"ärztinte",
-			"ärztintet",
-			"ärztinten",
-			"ärztintest",
-			"ärztinete",
-			"ärztinetet",
-			"ärztineten",
-			"ärztinetest",
-			"ärztinend",
-			// Participle.
-			"geärztint",
-		] );
-	} );
-
-	it( "removes -e noun endings and adds -n/-ns noun endings for words ending in -e", () => {
-		expect( getForms( "tee", morphologyDataDE ) ).toEqual( [
-			"tee",
-			"tees",
-			// Irregular noun suffixes
-			"teen",
-			"teens",
-			// Adjective suffixes
-			"teee",
-			"teeem",
-			"teeen",
-			"teeer",
-			"teees",
-			"teer",
-			"teere",
-			"teerem",
-			"teeren",
-			"teerer",
-			"teeres",
-			"teeste",
-			"teestem",
-			"teesten",
-			"teester",
-			"teestes",
-			"teet",
-			"teest",
-			"teeest",
-			"teeet",
-			"teete",
-			"teetet",
-			"teeten",
-			"teetest",
-			"teeete",
-			"teeetet",
-			"teeeten",
-			"teeetest",
-			"teeend",
-			// Participle.
-			"geteet",
-		] );
-	} );
-
-	it( "adds a form where -n is removed for nouns with a stem ending in -inn", () => {
-		expect( getForms( "ärztinnen", morphologyDataDE ) ).toEqual( [
-			// Regular noun suffixes.
-			"ärztinnen",
-			"ärztinne",
-			"ärztinnens",
-			"ärztinner",
-			"ärztinnern",
-			"ärztinners",
-			"ärztinnes",
-			"ärztinns",
-			// Adjective suffixes.
-			"ärztinnem",
-			"ärztinnere",
-			"ärztinnerem",
-			"ärztinneren",
-			"ärztinnerer",
-			"ärztinneres",
-			"ärztinnste",
-			"ärztinnstem",
-			"ärztinnsten",
-			"ärztinnster",
-			"ärztinnstes",
-			"ärztinnt",
-			"ärztinnst",
-			"ärztinnest",
-			"ärztinnet",
-			"ärztinnte",
-			"ärztinntet",
-			"ärztinnten",
-			"ärztinntest",
-			"ärztinnete",
-			"ärztinnetet",
-			"ärztinneten",
-			"ärztinnetest",
-			"ärztinnend",
-			// Participle.
-			"geärztinnt",
-			// Stem
-			"ärztinn",
-			// Additional noun suffix.
-			"ärztin",
-		] );
 	} );
 
 	it( "returns all relevant forms if a stem is on an exception list in more than one category, e.g." +
