+++ packages/yoastseo/src/morphology/german/generateRegularVerbForms.js
@@ -5,7 +5,7 @@ import { generateParticipleForm } from "./generateParticipleForm";
  * Generates regular verb forms.
  *
  * @param {Object}  morphologyDataVerbs The German morphology data for verbs.
- * @param {string}  stemmedWord         The stemmed word for which to create the past participle form.
+ * @param {string}  stemmedWord         The stemmed word for which to create the regular verb forms.
  *
  * @returns {string[]} The created verb forms.
  */
