+++ packages/yoastseo/src/morphology/german/getForms.js
@@ -3,100 +3,11 @@ import { detectAndStemRegularParticiple } from "./detectAndStemRegularParticiple
 import { generateAdjectiveExceptionForms } from "./generateAdjectiveExceptionForms";
 import { generateNounExceptionForms } from "./generateNounExceptionForms";
 import { generateRegularVerbForms } from "./generateRegularVerbForms";
+import { generateRegularNounForms } from "./generateRegularNounForms";
 import { generateVerbExceptionForms } from "./generateVerbExceptionForms";
 import stem from "./stem";
 
 import { uniq as unique } from "lodash-es";
-
-/**
- * Adds suffixes to the list of regular suffixes.
- *
- * @param {Object}          morphologyDataSuffixAdditions   The German data for suffix additions.
- * @param {Array<string>}   regularSuffixes                 All regular suffixes for German.
- * @param {string}          stemmedWordToCheck              The stem to check.
- *
- * @returns {Array<string>} The modified list of regular suffixes.
- */
-const addSuffixesToRegulars = function( morphologyDataSuffixAdditions, regularSuffixes, stemmedWordToCheck ) {
-	for ( const key of Object.keys( morphologyDataSuffixAdditions ) ) {
-		const endingsToCheck = morphologyDataSuffixAdditions[ key ][ 0 ];
-		const suffixesToAdd = morphologyDataSuffixAdditions[ key ][ 1 ];
-
-		// Append to the regular suffixes if one of the endings match.
-		if ( endingsToCheck.some( ending => stemmedWordToCheck.endsWith( ending ) ) ) {
-			regularSuffixes = regularSuffixes.concat( suffixesToAdd );
-		}
-	}
-
-	return regularSuffixes;
-};
-
-/**
- * Deletes suffixes from the list of regular suffixes.
- *
- * @param {Object}          morphologyDataSuffixDeletions   The German data for suffix deletions.
- * @param {Array<string>}   regularSuffixes                 All regular suffixes for German.
- * @param {string}          stemmedWordToCheck              The stem to check.
- *
- * @returns {Array<string>} The modified list of regular suffixes.
- */
-const removeSuffixesFromRegulars = function( morphologyDataSuffixDeletions, regularSuffixes, stemmedWordToCheck ) {
-	for ( const key of Object.keys( morphologyDataSuffixDeletions ) ) {
-		const endingsToCheck = morphologyDataSuffixDeletions[ key ][ 0 ];
-		const suffixesToDelete = morphologyDataSuffixDeletions[ key ][ 1 ];
-
-		// Delete from the regular suffixes if one of the endings match.
-		if ( endingsToCheck.some( ending => stemmedWordToCheck.endsWith( ending ) ) ) {
-			regularSuffixes = regularSuffixes.filter( ending => ! suffixesToDelete.includes( ending ) );
-		}
-	}
-
-	return regularSuffixes;
-};
-
-/**
- * Adds or removes suffixes from the list of regulars depending on the ending of the stem checked.
- *
- * @param {Object}          morphologyDataNouns The German morphology data for nouns.
- * @param {Array<string>}   regularSuffixes     All regular suffixes for German.
- * @param {string}          stemmedWordToCheck  The stem to check.
- *
- * @returns {Array<string>} The modified list of regular suffixes.
- */
-const modifyListOfRegularSuffixes = function( morphologyDataNouns, regularSuffixes, stemmedWordToCheck ) {
-	const additions = morphologyDataNouns.regularSuffixAdditions;
-	const deletions = morphologyDataNouns.regularSuffixDeletions;
-
-	regularSuffixes = addSuffixesToRegulars( additions, regularSuffixes, stemmedWordToCheck );
-	regularSuffixes = removeSuffixesFromRegulars( deletions, regularSuffixes, stemmedWordToCheck );
-
-	return regularSuffixes;
-};
-
-/**
- * Add forms based on changes other than simple suffix concatenations.
- *
- * @param {Object}  morphologyDataNouns The German morphology data for nouns.
- * @param {string}  stemmedWordToCheck  The stem to check.
- *
- * @returns {Array<string>} The modified forms.
- */
-const addFormsWithRemovedLetters = function( morphologyDataNouns, stemmedWordToCheck ) {
-	const forms = [];
-	const stemChanges = morphologyDataNouns.changeStem;
-
-	for ( const key of Object.keys( stemChanges ) ) {
-		const changeCategory = stemChanges[ key ];
-		const endingToCheck = changeCategory[ 0 ];
-
-		if ( stemmedWordToCheck.endsWith( endingToCheck ) ) {
-			const stemWithoutEnding = stemmedWordToCheck.slice( 0, stemmedWordToCheck.length - endingToCheck.length );
-			forms.push( stemWithoutEnding.concat( changeCategory[ 1 ] ) );
-		}
-	}
-
-	return forms;
-};
 
 /**
  * Creates morphological forms for a given German word.
@@ -144,13 +55,8 @@ export function getForms( word, morphologyData ) {
 		] );
 	}
 
-	// Modify regular suffixes assuming the word is a noun.
-	let regularNounSuffixes = morphologyData.nouns.regularSuffixes.slice();
-	// Depending on the specific ending of the stem, we can add/remove some suffixes from the list of regulars.
-	regularNounSuffixes = modifyListOfRegularSuffixes( morphologyData.nouns, regularNounSuffixes, stemmedWord );
-
 	// If the stem wasn't found on any exception list, add regular noun suffixes.
-	forms.push( ...regularNounSuffixes.map( suffix => stemmedWord.concat( suffix ) ) );
+	forms.push( ...generateRegularNounForms( morphologyData.nouns, stemmedWord ) );
 
 	// Also add regular adjective suffixes.
 	forms.push( ...addAllAdjectiveSuffixes( morphologyData.adjectives, stemmedWord ) );
@@ -160,12 +66,6 @@ export function getForms( word, morphologyData ) {
 
 	// Also add the stemmed word, since it might be a valid word form on its own.
 	forms.push( stemmedWord );
-
-	/*
-	 * In some cases, we need make changes to the stem that aren't simply concatenations (e.g. remove n from the stem
-	 * Ärztinn to obtain Ärztin.
-	 */
-	forms.push( ...addFormsWithRemovedLetters( morphologyData.nouns, stemmedWord ) );
 
 	return unique( forms );
 }
